# docs/swagger.yaml
definitions:
  dto.CreateTenantRequest:
    properties:
      document:
        type: string
      live:
        type: boolean
      name:
        type: string
    required:
    - document
    - live
    - name
    type: object
  dto.TenantResponse:
    properties:
      createAt:
        type: string
      document:
        type: string
      live:
        type: boolean
      name:
        type: string
      updateAt:
        type: string
      uuid:
        type: string
    type: object
  dto.TenantsResponse:
    properties:
      page:
        type: integer
      size:
        type: integer
      tenants:
        items:
          $ref: '#/definitions/dto.TenantResponse'
        type: array
    type: object
  rest_err.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_err.RestErr:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_err.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/tenant:
    get:
      description: Busca um tenant no sistema usando o UUID ou o Documento (CNPJ/CPF).
        Pelo menos um dos dois campos deve ser fornecido.
      parameters:
      - description: 'UUID do tenant a ser buscado. (Ex: 8871abf3-ed11-4770-b986-e8d98d022d4f)'
        in: query
        name: uuid
        type: string
      - description: 'Documento (CNPJ/CPF) do tenant a ser buscado. (Ex: 12345678901234)'
        in: query
        name: document
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tenant encontrado com sucesso.
          schema:
            $ref: '#/definitions/dto.TenantResponse'
        "400":
          description: Requisição inválida (UUID inválido, ou nenhum dos campos 'uuid'/'document'
            fornecido).
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "404":
          description: Tenant não encontrado com os dados fornecidos.
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Erro interno do servidor.
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Busca um Tenant
      tags:
      - Tenant
    post:
      consumes:
      - application/json
      description: Registra um novo tenant (empresa/organização) no sistema.
      parameters:
      - description: Objeto do tenant que precisa ser criado.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tenant criado com sucesso.
          schema:
            $ref: '#/definitions/dto.TenantResponse'
        "400":
          description: Requisição inválida (corpo JSON mal formatado ou dados de entrada
            inválidos).
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "409":
          description: Conflito (o 'document' fornecido já está em uso).
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Erro interno do servidor.
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Cria um novo Tenant
      tags:
      - Tenant
  /api/v1/tenant/list:
    get:
      description: Retorna uma lista paginada de todos os tenants registrados no sistema.
      parameters:
      - default: 1
        description: O número da página a ser retornada (deve ser >= 1).
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: O número de itens por página (máximo 100).
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de tenants retornada com sucesso.
          schema:
            $ref: '#/definitions/dto.TenantsResponse'
        "400":
          description: Requisição inválida (parâmetros de paginação ausentes ou inválidos,
            ou pageSize > 100).
          schema:
            $ref: '#/definitions/rest_err.RestErr'
        "500":
          description: Erro interno do servidor.
          schema:
            $ref: '#/definitions/rest_err.RestErr'
      summary: Lista Tenants
      tags:
      - Tenant
swagger: "2.0"